requirements.txt
youtube-transcript-api==0.6.2
openai==0.28.1
python-dotenv==1.0.1
requests==2.31.0 